backend/requirements.txt
backend==0.2.4.1
fastapi==0.115.12
openai==1.79.0
pydantic==2.11.4
python-dotenv==1.1.0
PyYAML==6.0.2
PyYAML==6.0.2
